+++ packages/stack/src/lib/stack-app.ts
@@ -83,11 +83,11 @@ function getUrls(partial: Partial<HandlerUrls>): HandlerUrls {
   const handler = partial.handler ?? "/handler";
   return {
     handler,
-    signIn: `${handler}/signin`,
+    signIn: `${handler}/sign-in`,
     afterSignIn: "/",
-    signUp: `${handler}/signup`,
+    signUp: `${handler}/sign-up`,
     afterSignUp: "/",
-    signOut: `${handler}/signout`,
+    signOut: `${handler}/sign-out`,
     afterSignOut: "/",
     emailVerification: `${handler}/email-verification`,
     passwordReset: `${handler}/password-reset`,
@@ -651,7 +651,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
       config: {
         credentialEnabled: crud.config.credential_enabled,
         magicLinkEnabled: crud.config.magic_link_enabled,
-        oauthProviders: crud.config.oauth_providers.map((p) => ({
+        oauthProviders: crud.config.enabled_oauth_providers.map((p) => ({
           id: p.id,
         })),
       }