+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,11 @@
 # e2e-tests
 
+## 2.5.5
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.5.5
+
 ## 2.5.4
 
 ### Patch Changes