+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @stackframe/docs
 
+## 2.5.5
+
 ## 2.5.4
 
 ### Patch Changes