+++ examples/cjs-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # cjs-test
 
+## 2.5.5
+
+### Patch Changes
+
+- @stackframe/stack@2.5.5
+
 ## 2.5.4
 
 ### Patch Changes