+++ packages/stack-sc/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @stackframe/stack-sc
 
+## 2.5.5
+
 ## 2.5.4
 
 ### Patch Changes