+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/stack
 
+## 2.5.5
+
+### Patch Changes
+
+- @stackframe/stack-sc@2.5.5
+- @stackframe/stack-shared@2.5.5
+- @stackframe/stack-ui@2.5.5
+
 ## 2.5.4
 
 ### Patch Changes